\name{}
\alias{hello}
\title{Hello, World!}
\usage{
sort(x, decreasing = FALSE, ...)

## Default method:
sort(x, decreasing = FALSE, na.last = NA, ...)

sort.int(x, partial = NULL, na.last = NA, decreasing = FALSE,
         method = c("auto", "shell", "quick", "merge"), index.return = FALSE)
}
\arguments{

}
\description{
Prints 'Hello, world!'.
}
\examples{
hello()
}
